requirements.txt
<<<<<<< HEAD
# Core FastAPI dependencies
fastapi==0.119.0
uvicorn[standard]==0.25.0
pydantic==2.11.7
python-multipart==0.0.20

# HTTP and API clients
requests==2.32.5
httpx==0.28.1

# GitHub integration
PyGithub==2.8.1

# OpenAI integration
openai==1.99.9

# Environment and configuration
python-dotenv==1.1.1

# Logging and monitoring
python-json-logger==3.3.0

# Security and validation
email-validator==2.3.0

# Async support
aiofiles==23.2.1

# Development and testing (optional)
pytest==8.4.2
pytest-asyncio==0.23.8
black==25.1.0
flake8==7.3.0

# Database support (for future use)
# sqlalchemy==2.0.43
# psycopg2-binary==2.9.10
# alembic==1.16.5

# Redis support (for future use)
# redis==5.1.0
# celery==5.3.6
=======
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
boto3==1.40.30
botocore==1.40.30
cachetools==5.5.2
caer==2.0.8
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cmake==3.31.4
colorama==0.4.6
comm==0.2.3
contourpy==1.2.1
cryptography==45.0.7
cycler==0.12.1
debugpy==1.8.16
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.0
fastapi==0.119.0
fastjsonschema==2.21.1
fastuuid==0.12.0
feedparser==6.0.12
ffmpy==0.5.0
filelock==3.19.1
filetype==1.2.0
fire==0.7.0
flake8==7.3.0
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.5.4
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.38.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==5.20.0
gradio_client==1.7.2
greenlet==3.2.4
groovy==0.1.2
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.11.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
isort==6.0.1
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.4.2
jq==1.10.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.3.3
kiwisolver==1.4.5
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
litellm==1.77.1
llvmlite==0.43.0
lxml==6.0.2
Mako==1.3.10
Markdown==3.6
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.0
motor==3.3.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.6.4
mypy==1.18.1
mypy_extensions==1.1.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook==7.4.5
notebook_shim==0.2.4
numba==0.60.0
numpy==2.3.3
oauthlib==3.3.1
onnx==1.17.0
openai==1.99.9
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.15
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
plotly==5.22.0
pluggy==1.6.0
pooch==1.8.2
postgrest==2.19.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybboxes==0.1.6
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.4.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PyNaCl==1.6.0
pyparsing==3.2.4
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.1
realtime==2.19.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
roboflow==1.1.54
rpds-py==0.27.1
rsa==4.9.1
ruff==0.9.9
s3transfer==0.14.0
s5cmd==0.2.0
safehttpx==0.1.6
sahi==0.11.20
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==70.0.0
sgmllib3k==1.0.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
storage3==2.19.0
StrEnum==0.4.15
stripe==12.5.1
supabase==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.1
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.4
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.82
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xlrd==2.0.1
yarl==1.20.1
yolov5==7.0.14
youtube-dl==2021.12.17
zipp==3.23.0
>>>>>>> 3a44616399e0ef6af1c4d0d7d4e7679041dbee90
